test: 발자국 API 문서화 클래스를 포맷에 맞게 수정
